requirements.txt
asgiref==3.3.1
azure-core==1.10.0
azure-identity==1.5.0
bleach==3.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
coverage==5.3.1
cryptography==3.3.1
Django==3.1.5
docutils==0.16
idna==2.10
keyring==22.0.0
msal==1.8.0
msal-extensions==0.3.0
packaging==20.8
pkginfo==1.7.0
portalocker==1.7.1
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.5
readme-renderer==28.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
sqlparse==0.4.1
tqdm==4.56.0
twine==3.3.0
urllib3==1.26.2
webencodings==0.5.1
